requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
opencv-python==4.5.3.56
mediapipe==0.8.9.1
requests==2.26.0
pyserial==3.5 